requirements.txt
arrow==1.3.0
asgiref==3.9.0
binaryornot==0.4.4
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
dj-database-url==3.0.1
Django==5.2.4
django-browser-reload==1.18.0
django-debug-toolbar==6.0.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
